6D-pose-estimation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argon2-cffi=20.1.0=py38h25fe258_2
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.2.1=pyh9f0ad1d_0
ca-certificates=2020.12.8=h06a4308_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py38h06a4308_0
cffi=1.14.0=py38h2e261b9_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.2.10=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
fribidi=1.0.10=h7b6447c_0
git=2.23.0=pl526hacde149_0
glib=2.63.1=h5a9c865_0
graphite2=1.3.14=h23475e2_0
graphviz=2.40.1=h21bd128_2
harfbuzz=1.8.8=hffaf4a1_0
hiddenlayer=0.3=pypi_0
icu=58.2=he6710b0_3
importlib-metadata=3.1.1=pyhd8ed1ab_0
importlib_metadata=3.1.1=hd8ed1ab_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h81c977d_1
ipython=7.19.0=py38h81c977d_0
ipython_genutils=0.2.0=py_1
jedi=0.17.2=py38h578d9bd_1
jinja2=2.11.2=pyh9f0ad1d_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py38h578d9bd_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
kiwisolver=1.3.1=pypi_0
krb5=1.18.2=h173b8e3_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcurl=7.71.1=h20c2e04_1
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
markupsafe=1.1.1=py38h8df0ef7_2
matplotlib=3.1.2=pypi_0
mistune=0.8.4=py38h25fe258_1002
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38h578d9bd_3
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pyhd8ed1ab_0
ninja=1.10.2=py38hff7bd54_0
notebook=6.1.5=py38h578d9bd_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
opencv-contrib-python=4.2.0.34=pypi_0
openssl=1.1.1i=h27cfd23_0
packaging=20.7=pyhd3deb0d_0
pandoc=2.11.2=h36c2ea0_0
pandocfilters=1.4.2=py_1
pango=1.42.4=h049681c_0
parso=0.7.1=pyh9f0ad1d_0
pcre=8.44=he6710b0_0
perl=5.26.2=h14c3975_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.0.1=py38he98fc37_0
pip=20.3.1=py38h06a4308_0
pixman=0.40.0=h7b6447c_0
prometheus_client=0.9.0=pyhd3deb0d_0
prompt-toolkit=3.0.8=pyha770c72_0
ptyprocess=0.6.0=py_1001
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.3=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py38h25fe258_1
python=3.8.2=hcf32534_0
python-dateutil=2.8.1=py_0
python-graphviz=0.15=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.5.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pyzmq=20.0.0=py38h1d1b12f_1
readline=8.0=h7b6447c_0
scipy=1.4.1=pypi_0
send2trash=1.5.0=py_0
setuptools=51.0.0=py38h06a4308_2
six=1.15.0=py38h06a4308_0
sqlite=3.33.0=h62c20be_0
terminado=0.9.1=py38h32f6830_1
testpath=0.4.4=py_0
tk=8.6.10=hbc83047_0
torchsummary=1.5.1=pypi_0
torchvision=0.6.0=py38_cu101
tornado=6.1=py38h25fe258_0
traitlets=5.0.5=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.36.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=h58526e2_3
zipp=3.4.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
